pythonProject2/requirements.txt
numpy
pandas
streamlit
altair
plotly_express
unidecode
